refactor: remove cert package, rename client package to cmd
